feat: solve task

+++ src/resources/users/user.router.js
@@ -1,5 +1,4 @@
 const router = require('express').Router();
-const { v4: uuidv4 } = require('uuid');
 const User = require('./user.model');
 const usersService = require('./user.service');
 const { StatusCode } = require('../../common/constants');
@@ -13,13 +12,11 @@ router.route('/')
     .json(users.map(User.toResponse));
   })
   .post( async (req, res) => {
-    const newUser = req.body;
-    newUser.id = uuidv4();
-    await usersService.createUser(newUser);
+    const UserBody = req.body;
+    const newUser = await usersService.createUser(UserBody);
     res
       .status(StatusCode.Created)
       .json(User.toResponse(newUser));
-
 });
 
 router.route('/:id')
@@ -29,24 +26,39 @@ router.route('/:id')
       res
         .status(StatusCode.BadRequest)
         .json({error: `User id = ${userId} is not valid`});
+    } else {
+      const user = await usersService.getUserBy(userId);
+      if (!user) {
+        res
+        .status(StatusCode.NotFound)
+        .json({error: `User id = ${userId} is not found`});
+      } else {
+        res
+        .status(StatusCode.OK)
+        .json(User.toResponse(user));
+      }
     }
-    const user = await usersService.getUserBy(userId);
-    res
-      .status(StatusCode.OK)
-      .json(User.toResponse(user));
   })
   .put( async (req, res) => {
     const userId = req.params.id;
-    const newUser = req.body;
+    const UserBody = req.body;
     if (!checkvalidityUIID(userId)) {
       res
         .status(StatusCode.BadRequest)
         .json({error: `User id = ${userId} is not valid`});
+    } else {
+      const updatedUser = await usersService.updateUserBy(userId, UserBody);
+      if (!updatedUser) {
+        res
+        .status(StatusCode.NotFound)
+        .json({error: `User id = ${userId} is not found`});
+      } else {
+        res
+        .status(StatusCode.OK)
+        .json(User.toResponse(updatedUser));
+      }
     }
-    const user = await usersService.putUserBy(userId, newUser);
-    res
-      .status(StatusCode.OK)
-      .json(User.toResponse(user));
+
   })
   .delete( async (req, res) => {
     const userId = req.params.id;
@@ -54,11 +66,19 @@ router.route('/:id')
       res
         .status(StatusCode.BadRequest)
         .json({error: `User id = ${userId} is not valid`});
+    } else {
+      const user = await usersService.getUserBy(userId);
+      if (!user) {
+        res
+        .status(StatusCode.NotFound)
+        .json({error: `User id = ${userId} is not found`});
+      } else {
+        await usersService.deleteUserBy(userId);
+        res
+          .status(StatusCode.NoContent)
+          .end();
+      }
     }
-    await usersService.deleteUserBy(userId);
-    res
-      .status(StatusCode.NoContent)
-      .end();
 });
 
 module.exports = router;
